requirements.txt
pip                22.3.1
mysql-connector    2.2.9